Rename PanelView to IndividualResultsTableTablet
