+++ packages/esm-patient-tests-app/src/test-results/individual-results-table-tablet/individual-results-table-tablet.scss
