+++ packages/esm-patient-tests-app/src/test-results/individual-results-table-tablet/lab-set-panel.scss
